Integrated Manager System
This commit includes a lot of changes because of the addition of several
project files that permit the use of managers and so on.  I still need
to fill it all in, but for now, its at least integrated.  I also forgot
to push the integrations first and started developing a home page, which
is included in the commit.